specify data structure for input services
